+++ waspc/data/Generator/templates/sdk/wasp/client/vite/plugins/validateEnv.ts
@@ -1,54 +1,67 @@
-import { type Plugin } from "vite";
-
+{{={= =}=}}
+import path from "node:path";
 import {
-  formatZodEnvError,
-  getValidatedEnvOrError,
-} from "../../../env/validation.js";
-import { colorize } from "../../../universal/ansiColors.js";
-import { getClientEnvSchema } from "../../env/schema.js";
-import { loadEnvVars } from "./envFile.js";
+  type Plugin,
+  type ResolvedConfig,
+  createServer as createViteServer,
+  isRunnableDevEnvironment
+} from "vite";
+
+const PLUGIN_NAME = "wasp:validate-env";
+const CLIENT_ENV_SCHEMA_VALIDATION_MODULE = "{= clientEnvSchemaValidationModulePath =}"
 
 export function validateEnv(): Plugin {
-  let validationResult: ReturnType<typeof getValidatedEnvOrError> | null = null;
+  let resolvedConfig: ResolvedConfig;
+
   return {
-    name: "wasp:validate-env",
-    async configResolved(config) {
-      const env = await loadEnvVars({
-        rootDir: config.root,
-        // We are sure that `envPrefix` is defined because
-        // we defined it in an earlier plugin.
-        envPrefix: config.envPrefix!,
-        // We load the env file variables only in development,
-        // when building for production, users are expected to
-        // provide the environment variables inline.
-        loadDotEnvFile: config.command === "serve",
-      });
-      const schema = getClientEnvSchema(config.mode);
-      validationResult = getValidatedEnvOrError(env, schema);
-
-      // Exit if we are in build mode, because we can't show the error in the browser.
-      if (config.command === "build" && !validationResult.success) {
-        const validationErrorMessage = formatZodEnvError(validationResult.error);
-        console.error(colorize("red", validationErrorMessage));
-        process.exit(1);
-      }
+    name: PLUGIN_NAME,
+    configResolved(config) {
+      resolvedConfig = config; 
     },
-    configureServer: (server) => {
-      if (validationResult === null || validationResult.success) {
-        return;
-      }
-
-      // Send the error to the browser.
-      const validationErrorMessage = formatZodEnvError(validationResult.error);
-      server.ws.on("connection", () => {
-        server.ws.send({
-          type: "error",
-          err: {
-            message: validationErrorMessage,
-            stack: "",
-          },
-        });
+    // We validate just before any artifacts are built.
+    async buildStart() {
+      // We need to import the client env schema validation module
+      // through a Vite server, because both the user and the Wasp schema
+      // modules may depend on bundler features.
+      // Because of that we spin up a tepomrary Vite server.
+      //
+      // Alternatively, for `serve`, we could use the Vite server provided
+      // through the `configureServer` hook, but that would complicate
+      // the solution for negligible performance benefits. 
+      const tempServer = await createViteServer({
+        root: resolvedConfig.root,
+        mode: resolvedConfig.mode,
+        // To ensure we pick up all user-defined plugins (resolution matches the main build)
+        // while avoiding recursion. This includes the `wasp` plugin.
+        configFile: false,
+        plugins: resolvedConfig.plugins.filter(
+          (plugin) => plugin.name !== PLUGIN_NAME
+        ),
+        // Minimize side effects from spinning up a temporary dev server.
+        appType: 'custom',      // avoid HTML handling
+        server: { 
+          middlewareMode: true, // do not start an actual HTTP server
+          watch: null, 
+          hmr: false 
+        },
+        logLevel: "silent",
+        optimizeDeps: { noDiscovery: true, include: [] },
+        clearScreen: false,
       });
+
+      try {
+        // Vite's `ssr` means bundled for "backend JS runtime", like Node.
+        // This envrionemnt is always runnable in vite dev server.
+        if (!isRunnableDevEnvironment(tempServer.environments.ssr)) {
+          throw new Error(`Expected ssr to be a runnable dev environment`)
+        }
+        // The imported module runs env schema validation as an import
+        // side-effect and throws on failure. 
+        const moduleAbsPath = path.resolve(resolvedConfig.root, CLIENT_ENV_SCHEMA_VALIDATION_MODULE);
+        await tempServer.environments.ssr.runner.import(moduleAbsPath);
+      } finally {
+        await tempServer.close();
+      }
     },
   };
 }